Ready to become installable package
