+++ odin-react/.OLD_CODE/src/components/LiveImage.jsx
