+++ odin-react/example/src/MainPage.js
@@ -8,7 +8,7 @@ import ToggleButtonGroup from 'react-bootstrap/ToggleButtonGroup';
 
 import {OdinCard} from 'odin-react';
 import StatusCard from './StatusCard'
-// import ControlCard from './ControlCard'
+import ControlCard from './ControlCard'
 // import LiveImage from '../../../src/components/LiveImage';
 import {StatusBox} from 'odin-react';
 // import { useOdinContext } from '../services/odinContext';
@@ -18,7 +18,7 @@ import {WithEndpoint} from 'odin-react';
 import React, { useEffect, useState, useContext } from 'react';
 
 const TestButton = WithEndpoint(ToggleButton);
-const testToggleGroup = WithEndpoint(ToggleButtonGroup);
+const TestToggleGroup = WithEndpoint(ToggleButtonGroup);
 
 function MainPage(props) {
   // const {cryoEndpoint, specEndpoint, acqEndpoint} = useOdinContext();
@@ -80,11 +80,11 @@ function MainPage(props) {
         </OdinCard>
       </Col>
       <Col>
-        {/* <ControlCard */}
-          {/* title="Basic Controls"
+        <ControlCard
+          title="Basic Controls"
           specEndpoint={specEndpoint}
-          acqEndpoint={acqEndpoint} */}
-        {/* /> */}
+          acqEndpoint={acqEndpoint}
+        />
       </Col>
     </Row>
     <Row>
@@ -92,7 +92,9 @@ function MainPage(props) {
       <TestButton id="testButton" value={false} endpoint={cryoEndpoint} fullpath="bakeout/enabled">Test</TestButton>
       </Col>
       <Col>
-      {/* <TestToggleGroup id="testGroup" type='radio' defaultValue={0} fullPath="" */}
+      <TestToggleGroup id="testGroup" type='radio' defaultValue={0} fullPath="">
+
+        </TestToggleGroup>
       </Col>
     </Row>
   </Container>
