+++ odin-react/src/components/ErrorBoundary.jsx
@@ -1,18 +1,15 @@
 import React from 'react';
 import Alert from 'react-bootstrap/Alert';
 class ErrorBoundary extends React.Component {
- constructor(props) {
- super(props);
- this.state = { error: null, info: null };
- }
- componentDidCatch(error, errorInfo) {
- // Catch errors in any components below and re-render with error message
- this.setState({
- error: error,
- errorInfo: errorInfo
- })
- // You can also log error messages to an error reporting service here
- }
+    constructor(props) {
+        super(props);
+            this.state = { error: null, info: null };
+        }
+    componentDidCatch(error, errorInfo) {
+    // Catch errors in any components below and re-render with error message
+    this.setState({error: error, errorInfo: errorInfo});
+    // You can also log error messages to an error reporting service here
+    }
  render() {
  if (this.state.errorInfo) {
  // Error path
