+++ odin-react/src/index.js
@@ -6,12 +6,13 @@ import styles from './styles.module.css'
 //   return <div className={styles.test}>Example Component: {text}</div>
 // }
 
-export { default as OdinNav } from './components/OdinNav';
+export { default as OdinApp } from './components/OdinApp';
 export { default as OdinCard } from './components/OdinCard';
 export { default as StatusBox } from './components/StatusBox';
 export { default as ToggleSwitch} from './components/ToggleSwitch';
 export { default as ErrorBoundary} from './components/ErrorBoundary';
 export { default as OdinForm} from './components/OdinForm';
+export { default as OdinToggleButtonGroup} from './components/OdinToggleButtons'
 
 export { default as useApiPut } from './services/useApiPut';
 export { default as usePeriodicFetch } from './services/usePeriodicFetch';
